Upgraded to 0.6

+++ partialCopy/pcp.py
@@ -19,12 +19,22 @@ def run(excuter,Wait=True,output=True):
 		return p.returncode
 	else:
 		return p
+
+def size_2_bytes(size):
+	units = {"B": 1, "K": 10 ** 3, "M": 10 ** 6, "G": 10 ** 9, "T": 10 ** 12, "P":10**15}
+	unit=size[-1]
+	if unit not in units:
+		return int(size)
+	number=size[:-1]
+	return int(float(number)*units[unit])
+
 def sizeof_fmt(num, suffix='B'):
-    for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
-        if abs(num) < 1024.0:
-            return "%3.1f%s%s" % (num, unit, suffix)
-        num /= 1024.0
-    return "%.1f%s%s" % (num, 'Yi', suffix)
+	for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
+		if abs(num) < 1024.0:
+			return "%3.1f%s%s" % (num, unit, suffix)
+		num /= 1024.0
+	return "%.1f%s%s" % (num, 'Yi', suffix)
+
 def run2(excuter,Wait=True):
 	PIPE=subprocess.PIPE
 	p=subprocess.Popen(excuter,stdout=PIPE,stderr=PIPE,shell=True)
@@ -68,48 +78,45 @@ def eprint(*args, **kwargs):
 
 def copy(src,dest,**kwargs):
 	print (kwargs)
-#if kwargs["dry_run"]:	 print("Running in Dry run mode")
 	doneFiles=[]
-	if src[-1]!="/": src+="/"
-	#logfile=kwargs.get("log",src+".pcp_log")
-	#if logfile==None: logfile=src+".pcp_log"
-	#lstfile=kwargs.get("lst",src+".pcp_lst")
-	#if lstfile==None: lstfile=src+".pcp_lst"
-	rsync_list=kwargs.get("rsync-list",src+".pcp_rsync_list/")
-	if rsync_list==None: rsync_list=src+".pcp_rsync_list/"
-	#if os.path.exists(logfile):
-	#	doneFolders=[l.strip() for l in open(logfile,'r').readlines()]
-	#	logFile=open(logfile,"a")
-	#else: logFile = open(logfile, "w")
+	partitions = {}
+	copy_list = []
+	badfiles = []
+	bigfiles = {}
 
-	#rsync_options=kwargs.get("rsync","")
-	#if rsync_options  == None: rsync_options=""
-	#if  kwargs["no_scan"] and os.path.exists(lstfile):
-	#	files=[l.strip() for l in open(lstfile,"r").readlines()]
-	#else:
+	if src[-1]!="/": src+="/"
+
+	rsync_list=kwargs.get("rsync-list",src+"pcp_rsync_list/")
+	if rsync_list==None: rsync_list=src+"pcp_rsync_list/"
+	if kwargs["new"] or kwargs["modified_after"]:
+		for file in os.listdir(rsync_list):
+			pf=[f.strip() for f in open(rsync_list+"/"+file).readlines()]
+			partitions[file.split(".")[-1]]={"files":pf[:],"remaining":0}
+			doneFiles.extend(pf[:])
 	print("Finding files to copy........")
 	fparams=kwargs.get("find_params",None) if kwargs.get("find_params",None) else ""
+	if kwargs["modified_after"]:
+		fparams += " -newermt %s"%(kwargs["modified_after"])
 	files=run("find  %s -type f  %s"%(src,fparams))
-	#	print(lstfile)
-	#	lst=open(lstfile,"w")
-	#	lst.write(files)
-	#	lst.close()
-	#	print("Wrote file list to %s"%lstfile)
+
 	files=files.split("\n")
 	print("Found %s files"%str(len(files)-1))
 
-
-	if not os.path.exists(dest): os.makedirs(dest)
+	if kwargs.get("dest",None):
+		if not os.path.exists(dest): os.makedirs(dest)
 	if not os.path.exists(rsync_list): os.makedirs(rsync_list)
-	dest_size=checkDest(dest)
-	dest_limit=dest_size*0.95
-	original_dest_limit=dest_limit
-	partitions={}
-	copy_list=[]
-	badfiles=[]
-	bigfiles={}
+	if not kwargs["dest_size"]:
+		dest_size=checkDest(dest)
+		dest_limit=dest_size*0.95
+		original_dest_limit=dest_limit
+	else:
+		dest_size=size_2_bytes(kwargs["dest_size"])
+		dest_limit=dest_size
+		original_dest_limit=dest_size
+
 	for file in files:
 		if file=="": continue
+		if file in doneFiles: continue
 		try:
 			fsize=float(os.path.getsize(file))
 		except:
@@ -132,49 +139,38 @@ def copy(src,dest,**kwargs):
 			if added: continue
 			partitions[len(partitions.keys())+1]={"files":copy_list,"remaining":dest_limit}
 			copy_list=[]
-#			if kwargs["dry_run"]:
 			dest_limit=original_dest_limit
 			copy_list.append(file)
 			dest_limit-=fsize
-#			else:
-#				print("HDD is full, breaking....")
-#				break
 	partitions[len(partitions.keys()) + 1] = {"files": copy_list, "remaining": dest_limit}
 	for key in partitions:
-		rl_name = rsync_list + ".pcp." + str(key)
+		rl_name = rsync_list + "pcp." + str(key)
+		if os.path.exists(rl_name) and not kwargs.get("force",None): continue
 		rl = open(rl_name, "w")
 		files_count = len(partitions[key]["files"])
 		rl.write("\n".join(partitions[key]["files"]))
 		rl.close()
 		print("%s saved to %s, remaining %s" % (files_count, rl_name, sizeof_fmt(partitions[key]["remaining"])))
-#	if not kwargs["dry_run"]:
-#		print (datetime.datetime.now())
-#		res='rsync --safe-links %s %s --files %s %s'%(Config.rsync_options,rsync_options,rsync_list[-1],savePath)
-#		print(res)
-#		code=run(res,output=False)
-#		print (datetime.datetime.now())
-#		if code==0:
-#			files=open(rsync_list[-1]).read()
-#			logFile.write(files)
-#			logFile.flush()
+
 	if len(badfiles)>0:
 		print ("The following files won't be copied")
 		print ("\n".join(["\t"+b for b in badfiles]))
 	if len(bigfiles.keys())>0:
 		print ("The following files are big and can't be  copied to the current media type")
-		print ("\n".join(["\t" +b for b in  bigfiles.keys()]))
+		print ("\n".join(["\t" +b for b in bigfiles.keys()]))
 		print ("Total Big files",sizeof_fmt(sum(bigfiles.values())))
 
 if __name__=="__main__":
 	import argparse
 	parser = argparse.ArgumentParser()
 	parser.add_argument("src", help="Source Directory")
-	parser.add_argument("dest", help="Dest Mountpoint")
-	#parser.add_argument("-lg", "--log", help="Log File to use")
-	#parser.add_argument("-ls", "--lst", help="List File to use")
-	parser.add_argument("-s","--save-to",help="Where to save rsync list")
-	#parser.add_argument("-ns", "--no-scan", help="Don't rescan the folder, this needs a previous run",action='store_true')
-	#parser.add_argument("-d", "--dry-run", help="don't run rsync, just break file list, assume equal size drives",action='store_true')
+	group = parser.add_mutually_exclusive_group()
+	group.add_argument("--dest", help="Destionation mountpoint")
+	group.add_argument("--dest-size", help="Destination size, given in bytes or using 1 letter unit B,K,M,G,T,P")
+	parser.add_argument("-s","--save-to",help="Where to save rsync list,default '$src/pcp_rsync_list/'")
+	parser.add_argument("-f","--force",help="Rewrite all lists again",action = 'store_true')
+	parser.add_argument("-n", "--new", help="Find New Files",action='store_true')
+	parser.add_argument("-ma", "--modified-after", help="Find files modified after certain time (YYYY-mm-dd)")
 	#parser.add_argument("-rs", "--rsync", help="Extra rsync parameters")
 	parser.add_argument("-fp", "--find-params", help="Parameters to find command")
 	args = parser.parse_args()
